refactor: #522 refactor React adaptation
Based on the changes made in Angular adaptation, make React adaptation look the
same.
