+++ packages/liferay-npm-bundler/src/steps/adapt/create-react-app/index.ts
@@ -3,23 +3,26 @@
  * SPDX-License-Identifier: LGPL-3.0-or-later
  */
 
-import {transformTextFile} from 'liferay-js-toolkit-core';
-
-import {buildBundlerDir, project} from '../../../globals';
+import {project} from '../../../globals';
 import * as log from '../../../log';
-import {copyFiles, findFiles} from '../../../util/files';
+import {findFiles} from '../../../util/files';
 import {runPkgJsonScript} from '../../../util/run';
 import {
+	copyStaticAssets,
 	processAdapterModules,
+	processCssFiles,
 	processPackageJson,
 	processWebpackBundles,
 } from '../index';
 import adaptStaticURLsAtRuntime from '../transform/js/operation/adaptStaticURLsAtRuntime';
-import tweakAttachmentToDOM from './transform/js/operation/tweakAttachmentToDOM';
-import replace from './transform/text/operation/replace';
+import tweakAttachmentToDOM from './tweakAttachmentToDOM';
+
+const assetGlobs = ['static/media/*'];
+const cssGlobs = ['static/css/*.css'];
+const jsGlobs = ['static/js/*.js'];
 
 /**
- * Run configured rules.
+ * Adapt create-react-app project
  */
 export default async function adaptCreateReactApp(): Promise<void> {
 	log.info(`Running React's build...`);
@@ -36,74 +39,23 @@ export default async function adaptCreateReactApp(): Promise<void> {
 
 	log.info('Copying static assets...');
 
-	copyStaticAssets();
+	await copyStaticAssets(assetGlobs);
 
 	log.info(`Processing CSS files...`);
 
-	await processCssFiles();
+	await processCssFiles(cssGlobs, assetGlobs);
 
 	log.info('Processing webpack bundles...');
 
 	await processWebpackBundles(
+		jsGlobs,
 		tweakAttachmentToDOM(),
-		adaptStaticURLsAtRuntime('static/media/*')
+		adaptStaticURLsAtRuntime(...assetGlobs)
 	);
 
 	log.info('Processing package.json file...');
 
 	await processPackageJson(getMainCssPosixPath());
-}
-
-function copyStaticAssets(): void {
-	const copiedFiles = copyFiles(
-		project.dir.join(project.adapt.buildDir),
-		['static/media/*'],
-		buildBundlerDir
-	);
-
-	log.debug(`Copied ${copiedFiles.length} static assets`);
-}
-
-/**
- * Process CSS files to rewrite URLs to static assets so that they can be served
- * from Liferay.
- *
- * @remarks
- * This is a best effort approach that may not work when proxies or CDNs are
- * configured because we are hardcoding the '/o' part of the URL and not using
- * the adapt runtime to rewrite the URLs.
- *
- * Of course that is because we cannot execute anything inside CSS files, so we
- * can only rewrite them at build time.
- */
-async function processCssFiles(): Promise<void> {
-	const adaptBuildDir = project.dir.join(project.adapt.buildDir);
-
-	const cssFiles = findFiles(adaptBuildDir, ['static/css/*.css']);
-
-	const assetURLsMap = findFiles(adaptBuildDir, ['static/media/*']).reduce(
-		(map, sourceAsset) => {
-			map.set(
-				sourceAsset.asPosix,
-				`o${project.jar.webContextPath}/${sourceAsset.asPosix}`
-			);
-
-			return map;
-		},
-		new Map<string, string>()
-	);
-
-	await Promise.all(
-		cssFiles.map(async (file) => {
-			await transformTextFile(
-				adaptBuildDir.join(file),
-				buildBundlerDir.join(file),
-				replace(assetURLsMap)
-			);
-		})
-	);
-
-	log.debug(`Processed ${cssFiles.length} CSS files`);
 }
 
 function getMainCssPosixPath(): string | undefined {
