+++ packages/liferay-npm-bundler/src/steps/adapt/create-react-app/tweakAttachmentToDOM.ts
